More spec fixes, delete a duplicate spec
